25/08

+++ admin/danhmuc.php
@@ -15,14 +15,61 @@ if (!isset($_SESSION['login'])) {
         <div class="container">
             <h2 class="mt-5 text-center ">Quản lí danh mục</h2>
             <div>
-                <a href="add-user.php" class="btn btn-primary mt-5 ">Thêm người danh mục</a>
+                <form class="row g-3 needs-validation" action="" method="POST">
+                    <div class="col-md-4">
+                        <label for="validationCustom01" class="form-label">Tên danh mục</label>
+                        <input type="text" class="form-control" name="tendanhmuc" placeholder="Tên danh muc" required>
+                    </div>
+                    <div class="col-md-4">
+                        <label for="validationCustom02" class="form-label">Mô tả</label>
+                        <input type="password" class="form-control" name="mota" placeholder="Mô tả" required>
+                    </div>
+                    <div class="col-md-2">
+                        <label for="validationCustom05" class="form-label">Ngày đăng</label>
+                        <input type="text" class="form-control" name="ngaydang" placeholder="Ngày đăng" required>
+                    </div>
+                    <div class="col-md-2">
+                        <label for="validationCustom05" class="form-label">Quyền</label>
+                        <input type="text" class="form-control" name="dangnhap" required>
+                    </div>
+                    <div class="col-md-2">
+                        <label for="validationCustom05" class="form-label">Trạng thái</label>
+                        <select class="form-select" name="trangthai" required>
+                            <option selected disabled value="">Choose...</option>
+                            <option>Đã xét duyệt</option>
+                            <option>Chưa xét duyệt</option>
+                        </select>
+                    </div>
+                    <div class="col-12">
+                        <button class="btn btn-primary mt-3" type="submit" name="luuthongtin" value="luuthongtin">Thêm</button>
+                    </div>
+                </form>
+                <?php
+                if (isset($_POST['luuthongtin'])) {
+                    $tendanhmuc = $_POST['tendanhmuc'];
+                    $mota = $_POST['mota'];
+                    $ngaydang = $_POST['ngaydang'];
+                    $dangnhap = $_POST['dangnhap'];
+                    $trangthai = $_POST['trangthai'];
+
+                    $sql = "INSERT INTO danhmuc (tendanhmuc, mota, ngaydang, iddangnhap , idtrangthai)
+                        VALUES ('$tendanhmuc','$mota','$ngaydang',' $dangnhap','$trangthai')";
+                    // $result = mysqli_query($conn, $sql);
+                    // $count = mysqli_num_rows($result);
+                    if (mysqli_query($conn, $sql)) {
+                        header("Location:http://localhost:8080/btl/phenikaa/admin/danhmuc.php");
+                    } else {
+                        echo "Không thể thêm danh mục mới";
+                    }
+                }
+
+                ?>
             </div>
             <div class="row mt-5">
-                <table class="table table-bordered border-dark ">
+                <table class="table table-bordered border-dark table-striped table-hover">
                     <thead class="thead-dark">
                         <tr>
                             <th scope="col">TT</th>
-                            <th scope="col">Mã</th>
                             <th scope="col">Tên danh mục</th>
                             <th scope="col">Mô tả</th>
                             <th scope="col">Ngày đăng</th>
@@ -42,14 +89,12 @@ if (!isset($_SESSION['login'])) {
                         ?>
                                 <tr>
                                     <th scope="row"><?php echo $i; ?></th>
-                                    <td> <?php echo $row['iddanhmuc']; ?></td>
                                     <td> <?php echo $row['tendanhmuc']; ?> </td>
                                     <td> <?php echo $row['mota']; ?></td>
                                     <td> <?php echo $row['ngaydang']; ?></td>
                                     <td><a href="edit.php?id=<?php echo $row['id']; ?>"><i class="bi bi-pencil-square"></i></a></td>
-                                    <td><a href="delete.php?id=<?php echo $row['iddanhmuc']; ?>"
-                                    onclick="return confirm('Bạn có thực sự muốn xóa ?' );">
-                                    <i class="bi bi-archive-fill"></i></a></td>
+                                    <td><a href="delete.php?id=<?php echo $row['iddanhmuc']; ?>" onclick="return confirm('Bạn có thực sự muốn xóa ?' );">
+                                            <i class="bi bi-archive-fill"></i></a></td>
                                 </tr>
                         <?php
                                 $i++;
